Final version of direction cipher.
Thanks to accelarator, it was possible to implement a logic that create
direction puzzle that are necessary to open a gate. Code provides logic
to operate with adxl345 and guess in which direction our dice/cube is
heading to. Using special combination of cube/dice side placements,
person is able to unlock it.